feature: Initial physics prototype
It's very ugly and unoptimized right now. I made this
to figure out how to even approach making physics and
what problems I may have.

The main issue right now is that the particle coordinates
are in integers, which loses a lot of detail. The solution
for this is to increase the internal field from 800x600 to
something like 8000x6000 and go from there. This is what I'm
planning to do next, because currently the movement of the particles
is honestly awful.